[tcms-api] Leave only transport classes
this removes most of the top-level/oop interface behind the API
client. Because it is legacy code we feel like we should start
from scratch and implement OOP helpers if requested or when the
underlying server response becomes unreadable.

The resulting code should be easier to test and maintain.

Also remove tcms_api/tests/

the remaining tests are calling server side XML-RPC methods which
is also done as part of testing the xmlrpc/api from the Django
application. No need to repeat the same kind of testing!